Fingering on notes with beams did not respect min distance to staff
When a note is beamed to another note (e.g. two 8th notes) and the stems are pointing up and the note has a fingering, the fingering is positioned a min distance away from the beam but not the staff so it can end up being less than min distance away from the staff.