Allow the F2 key to cycle through the grips/handles of a shape (e.g. slur/tie) just like the TAB key
Problem solved: when adjusting a handle of a shape, such as slur or tie, the presence of the handles does not allow the user to judge how well the shape has been adjusted. So they have to hide the handles and if they are unhappy with the result, they have to display the handles again, activate the same handle and continue moving it. Using the keyboard this can be done by pressing ESC to hide the handles (the shape stays selected) and then F2 to display them again (enter edit mode). So after pressing F2 to display the handles, now the user can continue pressing F2 to cycle through the handles instead of TAB (TAB will still work of course).