Added as setting for cycling through the grips/handles with the Edit Element shortcut key
Problem solved: Add a setting "UseEditElementKeyToCycleThroughGrips" which when set to true (default is false) will enable the user to cycle through the grips/handles of an element with the "Edit Element" key. The idea is that the user will be able to select the element with keyboard and press the "Edit element" shortcut key to put it in edit mode, i.e. display the grips/handles. It would then be more convenient to continue pressing the same key to activate the desired handle instead of pressing TAB (TAB will still work).