Added "ExportDialogWillFocusTheExportButton" setting to allow the Export button on the Export dialog to be initially focused
Problem solved: When the export dialog appears and the user wants to export immediately, using the keyboard they have to navigate to the Export button and click Space or Enter. This is cumbersome so this change introduces a new setting called "ExportDialogWillFocusTheExportButton" which when true will make the Export button have the focus when the Export dialog appears.