added the "FixLostTrillColorOnScoreOpen" setting
When the "FixLostTrillColorOnScoreOpen" setting is true, this will fix an issue where a trill color is saved properly but not respected when the score is opened. The trill must have all of its properties set to the default values (offset, min distance, etc.). This gives a workaround: alter a property only slightly, e.g. if default min distance is 0.5, make it 0.51. This will make the trill segment go into the score and is then properly loaded. If all properties have their default values, the trill segment is not saved into the score and is recreated when the score is opened and there is a glitch there that doesn't transfer the color to the created segment properly.