Speed up the open of the palette search
When opening the palette search and there was no search text, all of the palettes and palette items were returned (no filtering was performed). Although the search results were invisible, this was a time consuming operation and was causing the search to open with a significant delay. This changeset causes no results to be returned when there is no search text thus speeding things up significantly. As a side effect, slows down the search when the first search character is typed because the search results are empty and have to be filled out (i.e. lots of elements to add to the listview).