Speed up palette search by having one right-click mouse area instance
This changeset creates one MouseArea for the entire palette GridView to listen for right-clicks. On reception of a right-click, the context menu will be bound to the grid view item under the mouse and displayed. Not having a separate right-click MouseArea per grid view item should further speed things up.